% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{add_human_homologs_to_mice_df}
\alias{add_human_homologs_to_mice_df}
\title{This function will process a mice cell type markers df to to keep only genes with one-to-one human orthology and 
will add the corresponding human homolog to each row. This will add human homologs in order of descending log2FC.}
\usage{
add_human_homologs_to_mice_df(cell_type_df, human_homologs_df)
}
\arguments{
\item{cell_type_df}{A data.frame with a list of markers for a cell type. The column name for the gene symbols has to be "Gene_symbol".}

\item{human_homologs_df}{A data.frame of mouse-human homologs output by covert_mouse_to_human()}
}
\value{
A data.frame with one-to-one orthologs and the corresponding human homologs.
}
\description{
This function will process a mice cell type markers df to to keep only genes with one-to-one human orthology and 
will add the corresponding human homolog to each row. This will add human homologs in order of descending log2FC.
}
