% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{calc_gene_cors}
\alias{calc_gene_cors}
\title{Correlation of expression for a target gene across the entire transcriptome.}
\usage{
calc_gene_cors(cell_types, metadata_df, exp_matrix, target_gene, cor_method)
}
\arguments{
\item{cell_types}{A character vector indicating cell names of interest.}

\item{metadata_df}{A data.frame with the metadata for the counts matrix.}

\item{exp_matrix}{A genes x cells counts matrix.}

\item{target_gene}{A string indicating the target gene to correlate with the rest of the transcriptome.}

\item{cor_method}{A string indicating the correlation method to use, either spearman or pearson}
}
\value{
A data.frame with the correlation of the target gene with all other genes in the count matrix.
}
\description{
This function will calculate the correlation between a target gene and other genes in a specific group of cells.
}
