% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calcCentroidDist}
\alias{calcCentroidDist}
\title{Compute centroid distances.}
\usage{
calcCentroidDist(seuratObj, embeddingKey, groupLabels, targetLabel)
}
\arguments{
\item{embeddingKey}{A character indicating the name of the embedding used
for comuting centroids.}

\item{groupLabels}{A character indicating the name of the cell type annotation
across which to compute centroids.}

\item{targetLabel}{the target annotation that we want to comapre other
annotations with. Euclidean distance of target annotation to itself should
be zero of course.}

\item{seuratobj}{A Seurat object.}
}
\value{
A dataframe with euclidean distance of a target cell annotation centroid
 to the centroids of other clusters or annotations.
}
\description{
Find centroids and compute distance of a target annotation centroid 
to centroids of other clusters.
}
