% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{calcGeneCors}
\alias{calcGeneCors}
\title{Correlation of expression for a target gene across the entire transcriptome.}
\usage{
calcGeneCors(
  seuratObj,
  assay = "SCT",
  targetGene,
  corMethod = "pearson",
  annoVariable = NULL,
  cellTypes = NULL
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{assay}{A character vector indicating the assay from which to get the counts.}

\item{targetGene}{A string indicating the target gene to correlate with the rest of the transcriptome.}

\item{corMethod}{A string indicating the correlation method to use, either spearman or pearson}

\item{annoVariable}{A character vector indicating the name of the metadata column that has the cell type annotations}

\item{cellTypes}{A character vector indicating the cell type annotations in which to calculate the correlations}
}
\value{
A data.frame with the correlation of the target gene with all other genes in the count matrix.
}
\description{
This function will calculate the correlation between a target gene and other genes in a specific group of cells.
}
\examples{
\dontrun{
calcGeneCors(seuratObj, targetGene = "MYH11")
}

}
