% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{calcSilscores}
\alias{calcSilscores}
\title{Calculation of silhouette scores}
\usage{
calcSilscores(seuratObj, clusteringRes, clusteringAlg = "louvain")
}
\arguments{
\item{seuratObj}{A Seurat obj with reduced dims embedding like pca. Default now is set to PCA.}

\item{clusteringRes}{A Numeric vector with clustering resolutions to test}
}
\value{
A Data.frame with silhouette scores for a the provided clustering resolutions
}
\description{
Helper function to calculate clusters silhouette scores from a seurat object.
}
\examples{
\dontrun{
sil_scores = calc_sil_scores(seurat_obj, seq(0.3, 0.9, by=0.1))
}
}
