% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{convert_mouse_to_human}
\alias{convert_mouse_to_human}
\title{Human and mouse homologs}
\usage{
convert_mouse_to_human(gene_vec, mouse_biomart)
}
\arguments{
\item{gene_vec}{A vector of mice genes}

\item{mouse_biomart}{A mouse biomart db from which gene attributes will be extracted}
}
\value{
A List where the first item is a data.frame object with 
mice gene names, corresponding human homologs, chromosome and type of orthology.
Second obj in the list is a vector with the names of genes that have no human homolog and
thus could not be converted.
}
\description{
This is a handy function to gene human homologs for mice genes.
}
