% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{custom_theme}
\alias{custom_theme}
\title{Custom theme to make ggplots prettier}
\usage{
custom_theme()
}
\value{
A ggplot theme
}
\description{
I'm not a big fan of the standard way Seurat ouputs UMAP plots
so this function will provide a proper aspect ratio and also
decently sized x and y axis label.
}
