% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{decontX_remove}
\alias{decontX_remove}
\title{Removal of ambient RNA}
\usage{
decontX_remove(seurat_obj)
}
\arguments{
\item{seurat_obj:}{A seurat obj that has been filtered for doublets}
}
\value{
A seurat object with a decontaminated raw counts matrix
}
\description{
Internal decontX wrapper to standardize removal of ambient mRNA contamination
}
