% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{decontXremove}
\alias{decontXremove}
\title{Removal of ambient RNA.}
\usage{
decontXremove(seuratObj)
}
\arguments{
\item{seuratObj:}{A seurat obj that has been filtered for doublets}
}
\value{
A seurat object with a decontaminated raw counts matrix and 
a metadata column showing the contamination scores output by decontX.
}
\description{
Internal decontX wrapper to standardize removal of ambient mRNA contamination
}
