% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{eucDist}
\alias{eucDist}
\title{Euclidean distances.}
\usage{
eucDist(coord1, coord2)
}
\arguments{
\item{coord1}{A 2 element numeric vector with coordinates for point 1.}

\item{coord2}{A 2 element numeric vector with coordinates for point 2.}
}
\value{
The euclidean distance between the two given coordinates.
}
\description{
Compute Euclidean distances between two points. X and Y coordinates
must be given for each point.
}
