% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{findClusterCentroids}
\alias{findClusterCentroids}
\title{Find cluster centroids.}
\usage{
findClusterCentroids(seuratObj, embeddingKey, groupLabels, dims = c(1, 2))
}
\arguments{
\item{seuratObj}{A Seurat object with dim reduced embeddings.}

\item{embeddingKey}{A character indicating the name of the embedding (e.g., pca).}

\item{groupLabels}{A character indicating the label by which to group the
cells for computing the centroids.}

\item{dims}{A numeric vector with the dimensions of the embedding across which we
wish to compute the cluster centroids. Default is to use only the the first 2 dims.}
}
\value{
A dataframe with centroids computed across each PC or embedding 
 column vector.
}
\description{
Compute the cluster centroids for a given set of cell type labels. They could
be cell type annotation or any cluster identification number.
}
