% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_human_ensembl_ids}
\alias{get_human_ensembl_ids}
\title{This function will convert gene symbols to Ensembl IDs}
\usage{
get_human_ensembl_ids(gene_symbols, human_biomart)
}
\arguments{
\item{gene_symbols}{A character vector with gene symbols to convert to Ensembl IDs.}

\item{human_biomart}{A human biomart object.}
}
\value{
A data.frame with gene symbols and their respective Ensembl IDs.
}
\description{
This function will convert gene symbols to Ensembl IDs
}
