% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{gprofiler_bar_plot}
\alias{gprofiler_bar_plot}
\title{This is a function to plot gene set enrichment analyses from gProfiler2
gprofiler_out: A list containing gProfiler2 output grom gost function}
\usage{
gprofiler_bar_plot(gprofiler_df, min_n_genes, max_n_genes, pathway_db, n_hits)
}
\arguments{
\item{min_n_genes}{A numeric indicating Min number of genes in gene set}

\item{max_n_genes}{A numeric indicating max number of genes in gene set}

\item{pathway_db}{A character indicating the name of the pathway db to be tested for enrichment}

\item{n_hits}{A numeric indicating how many terms should be plotted. Terms are ordered in descending FDR}
}
\value{
A ggplot object
}
\description{
This is a function to plot gene set enrichment analyses from gProfiler2
gprofiler_out: A list containing gProfiler2 output grom gost function
}
