% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{gprofiler_go_simplify}
\alias{gprofiler_go_simplify}
\title{This function will take a df output with GO terms by gProfiler enrichment analysis
and remove GO redundant terms to get a clearer picture of relevant biological processes.}
\usage{
gprofiler_go_simplify(
  gprofiler_res_df,
  cutoff = 0.7,
  select_fun = min,
  measure = "Rel",
  ontology,
  semData
)
}
\arguments{
\item{gprofiler_res_df}{A data.frame output by gProfiler gost() function. This should be the second item in the output list.}

\item{cutoff}{A numeric indicating the cutoff for semantic similarity}

\item{select_fun}{A function to select one term to keep from a subset of redundant terms based on p values.}

\item{measure}{A character indicating the semantic simlarity metric to be used.}

\item{ontology}{A character indicating which GO ontology to use (e.g., "BP")}

\item{semData}{The output of GoSemSim goData function that will allow comparison of semantic similarity between 2 terms.}
}
\value{
A vector with redundant GO terms meeting the similarity cutoff that should be removed.
}
\description{
This function will take a df output with GO terms by gProfiler enrichment analysis
and remove GO redundant terms to get a clearer picture of relevant biological processes.
}
