% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{makeSumStatsDF}
\alias{makeSumStatsDF}
\title{Helper function to make a summary df from metadata vars}
\usage{
makeSumStatsDF(seuratObj, seqPlatform)
}
\arguments{
\item{seuratObj}{A Seurat object with nUMI, nFeatures and other QC metrics}
}
\value{
A dataframe with summary statistics for several important QC metrics
}
\description{
This helper function will return summary stats for key QC metrics
as a dataframe.
}
