% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{maxAvgSil}
\alias{maxAvgSil}
\title{Internal helper function to get the max sil score from calc_sil_scores()}
\usage{
maxAvgSil(silScoresDf)
}
\arguments{
\item{silScoresDf}{A data.frame output by calc_sil_scores()}
}
\value{
A numeric indicating the maximum average sil sclore value across tested resolutions
}
\description{
Internal helper function to get the max sil score from calc_sil_scores()
}
