% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{miller_continous_scale}
\alias{miller_continous_scale}
\title{A nicer continuous scale than the default ggplot one.}
\usage{
miller_continous_scale(style = "points")
}
\arguments{
\item{style}{A character string indicating whether we need to plot points or bars.}
}
\value{
A ggplot continuous scale
}
\description{
A nicer continuous scale than the default ggplot one.
}
