% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{miller_discrete_scale}
\alias{miller_discrete_scale}
\title{A nicer discrete scale than the ggplot default one.}
\usage{
miller_discrete_scale(style = "points", option = 1)
}
\arguments{
\item{style}{A character string indicatin whether to plot points or bars.}

\item{option}{A numeric indicating whether to plot the first or second option.}
}
\value{
A ggplot discrete scale
}
\description{
A nicer discrete scale than the ggplot default one.
}
