% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot2GeneCor}
\alias{plot2GeneCor}
\title{Make a function to visualize the correlation between two genes.}
\usage{
plot2GeneCor(
  seuratObj,
  targetGenes,
  targetColData = NULL,
  targetAnno = NULL,
  showCor = FALSE,
  corMethod = "pearson",
  assay = "SCT"
)
}
\arguments{
\item{targetAnno}{A character vector with the names of the cell types in which we wish 
to correlate the genes.}

\item{showCor}{A boolean indicating whether we wish to calculate and show the correlation coefficient.}

\item{corMethod}{A character indicating the correlation method to use. One of "pearson" or "spearman".}

\item{assay}{A character indicating the assay from which to extract normalized counts. One of "SCT" or "RNA".}

\item{seurat_obj}{A seurat object with normalized expression values}

\item{targetColdata}{A character indicating the name of the metadata column with 
the cell types of interest.}
}
\value{
A ggplot object correlating expression of two genes in specific cell populations.
}
\description{
This function will take a seurat object as input and produce a dot plot 
correlating 2 genes in a specific group of cell types.
}
