% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_cors}
\alias{plot_cors}
\title{Make a function to plot correlations}
\usage{
plot_cors(cors_df, target_gene, target_cells, genes_to_label, point_size = 0.9)
}
\arguments{
\item{cors_df}{A data.frame that is output by calc_gene_cors()}

\item{target_gene}{A character string indicating the target gene}

\item{target_cells}{A character indicating the cells where the cors were calculated}

\item{genes_to_label}{A character vector with the names of the genes that should be highlighted in the plot.}

\item{point_size}{A numeric indicating the size of the points in the plot}
}
\value{
A ggplot object
}
\description{
This function will take the output of calc_gene_cors() and produce a
nice way to visualize the correlations.
}
