% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotExpressionOnPseudotime}
\alias{plotExpressionOnPseudotime}
\title{Model gene expression across pseudotime.}
\usage{
plotExpressionOnPseudotime(
  cdsOrdered,
  colDataVar,
  cellAnnotations,
  genes,
  facetWrap = FALSE,
  pseudotimeBoundary = 30
)
}
\arguments{
\item{cdsOrdered}{A monocle3 cds object with cells ordered across pseudotime.}

\item{colDataVar}{A character indicating the name of the column with the cell type annotations.}

\item{cellAnnotations}{A vector with cell types of interest for plotting gene expression.}

\item{genes}{A Vector with the genes to be plotted across pseudotime.}

\item{facetWrap}{A boolean indicating whether all genes shoud be shown in one or multiple plots.}

\item{pseudotimeBoundary}{A numeric indicating the upper boundary of pseudotime to be plotted.}
}
\value{
A ggplot object with one or multiple panels showing gene expression across pseudotime values.
}
\description{
This function will plot normalized expression values across pseudotime
}
