% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotFeatureUMAPList}
\alias{plotFeatureUMAPList}
\title{Named list of seurat gene expression plots}
\usage{
plotFeatureUMAPList(
  seuratObj,
  queryFeatures,
  targetAssay = "SCT",
  splitByVar = NULL,
  ptSize = 0.1
)
}
\arguments{
\item{seuratObj}{A seurat object with normalized expression values.}

\item{targetAssay}{A character indicating the assay to be used for making gene expression plots. 
Most of the time this argument will be either RNA or SCT. Default is set to SCT.}

\item{splitByVar}{A character indicating a metadat variable by which to facet the plot (e.g., sex, disease status).}

\item{ptSize}{A numeric indicating the size of the points for each plot. Default value is 0.1 but can be changed.}

\item{queryGenes}{A character vector with the names of the genes to be included in the list.}
}
\value{
A List of named ggplot objects, one for each of the genes included in the character vector.
}
\description{
A function that takes a seurat object as input and 
produces a named list of gene expression plots.
}
