% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotGene2TranscriptomeCor}
\alias{plotGene2TranscriptomeCor}
\title{Make a function to plot correlations of one gene to the entire trasncriptome}
\usage{
plotGene2TranscriptomeCor(
  corsDf,
  targetGene,
  targetCells,
  genesToLabel,
  pointSize = 0.9
)
}
\arguments{
\item{corsDf}{A data.frame that is output by \code{calcGeneCors}}

\item{targetGene}{A character string indicating the target gene}

\item{targetCells}{A character indicating the cells where the cors were calculated}

\item{genesToLabel}{A character vector with the names of the genes that should be highlighted in the plot.}

\item{pointSize}{A numeric indicating the size of the points in the plot}
}
\value{
A ggplot object
}
\description{
This function will take the output of calc_gene_cors() and produce a
nice way to visualize the correlations.
}
