% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotRNAcomplexityHist}
\alias{plotRNAcomplexityHist}
\title{Plot a histogram of RNA complexity scores}
\usage{
plotRNAcomplexityHist(seuratObj)
}
\arguments{
\item{seuratObj}{A Seurat obj with the variable
log10GenesPerUMI}
}
\value{
A ggplot object
}
\description{
This function will plot a histogram of 
RNA complexity scores
}
