% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotRNAFeature}
\alias{plotRNAFeature}
\title{Helper function to plot features}
\usage{
plotRNAFeature(seuratObj, feature)
}
\arguments{
\item{feature}{A character vector indicating the metadata variable 
to plot (i.e. nUMIs)}

\item{df}{A dataframe containing information about QC metrics.
This df should have a "libraryID" column.}
}
\value{
A ggplot object showing the distribution of QC metrics
across each cell in the library.
}
\description{
This function is equivalent to Seurat vlnPlot
but can be customized in any way we want to show
important QC metrics such as number of UMIs, detected
genes or percentage of reads mapped to mitochondrial 
or ribsomal genes.
}
