% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotRNAFeatureList}
\alias{plotRNAFeatureList}
\title{Plot features list}
\usage{
plotRNAFeatureList(seuratObj, features)
}
\arguments{
\item{seuratObj}{A Seurat object with metadata relevant to
QC metrics.}

\item{features}{A character vector with several QC metrics
to plot.}
}
\description{
Wrapper of `plotRNAFeatures` to show several features
at the same time
}
