% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotTopAvgExpr}
\alias{plotTopAvgExpr}
\title{Plot the most highly expressed genes in the dataset}
\usage{
plotTopAvgExpr(seuratObj, makeBoxplot = FALSE, nGenes = 50)
}
\arguments{
\item{seuratObj}{A Seurat object with the metadata variable "sample"}

\item{makeBoxplot}{A boolean indicating whether we want to generate a box plot.}

\item{nGenes}{A numeric indicating how many genes to plot. Default=50.}
}
\value{
A ggplot object
}
\description{
This function calculates the average expression of
highly variable genes in the data and then plots then
plots this value for the 50 most highly expressed.
}
