% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{scDblFinder_clusters}
\alias{scDblFinder_clusters}
\title{Identify and remove doublets.}
\usage{
scDblFinder_clusters(
  seurat_obj,
  nrep,
  multisample_dataset = FALSE,
  sample_ids = NULL
)
}
\arguments{
\item{nrep}{A Numeric indicating the number of times to run the scDblFinder function.}

\item{multisample_dataset}{A Logical indicating whether the processed dataset has multiple samples}

\item{sample_ids}{if multisample_dataset = TRUE, A character that refers to the metadata column
that contains sample ids.}

\item{seurat_obj:}{A Pre-processed seurat object with cell clusters annotations.}
}
\value{
A vector of consensus doublet cell barcodes
}
\description{
Internal helper function to automate doublet removal workflow and update metadata.
}
