% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{selectPCs}
\alias{selectPCs}
\title{Select n PCs}
\usage{
selectPCs(seuratObj)
}
\arguments{
\item{seuratObj}{A seurat obj with a computed DimReduc object}
}
\value{
A numeric indicating the number of PCs we need to explain the largest variation
}
\description{
This function will find the number of PCs explaining the largest
amount of variation in the data, which we need
for clustering and UMAP in a more data-driven way.
}
\examples{
\dontrun{
nPCs = selectPCs(seuratObj = seuratObj)
}
}
