% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{seurat_gene_plot_list}
\alias{seurat_gene_plot_list}
\title{Named list of seurat gene expression plots}
\usage{
seurat_gene_plot_list(
  seurat_object,
  target_assay = "SCT",
  genes_of_interest,
  split_by_var = NULL,
  pt.size = 0.1
)
}
\arguments{
\item{seurat_object}{A seurat object with normalized expression values.}

\item{target_assay}{A character indicating the assay to be used for making gene expression plots. 
Most of the time this argument will be either RNA or SCT. Default is set to SCT.}

\item{genes_of_interest}{A character vector with the names of the genes to be included in the list.}

\item{split_by_var}{A character indicating a metadat variable by which to facet the plot (e.g., sex, disease status).}

\item{pt.size}{A numeric indicating the size of the points for each plot. Default value is 0.1 but can be changed.}
}
\value{
A List of named ggplot objects, one for each of the genes included in the character vector.
}
\description{
A function that takes a seurat object as input and 
produces a named list of gene expression plots.
}
