% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{tidySeuratCounts}
\alias{tidySeuratCounts}
\title{Tidy Seurat counts into a data.frame with cell types of interest}
\usage{
tidySeuratCounts(
  seuratObj,
  assay = "SCT",
  targetColData = NULL,
  targetAnno = NULL
)
}
\arguments{
\item{assay}{A character indicating the name of the assay from which to extract the counts.
This can be set to SCT or RNA.}

\item{seurat_obj}{Seurat object from which we'll extract the normalized counts.}

\item{target_coldata}{A character indicating the name of the column with the cluster IDs or cell type annotations.}

\item{target_anno}{A character vector with the cluster IDs or names of cell types that will be included in the df.}
}
\value{
A data.frame of cells x genes with corresponding normalized counts. This df also includes a column with the 
cell barcode matching the cluster ID or cell type annotation.
}
\description{
Tidy Seurat counts into a data.frame with cell types of interest
}
