% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_processing_pipeline.R
\name{doItAll}
\alias{doItAll}
\title{Process scRNA libraries.}
\usage{
doItAll(
  countsMatrix,
  libraryID,
  studyID,
  tissueSource,
  seqPlatform = "10x",
  seqWorflow = "cell",
  UMI = TRUE,
  dblFindIter = 3,
  seuratFilter = TRUE,
  computeSil = TRUE,
  minRes = 0.2,
  maxRes = 0.8,
  clusteringAlg = "louvain",
  renameCells = FALSE,
  makeAnnData = FALSE,
  annDataParentDir = NULL,
  queryFeatures = NULL,
  ...
)
}
\arguments{
\item{countsMatrix}{sparse count matrix required as initial input. As of now, it accepts 
a filtered_feature_bc_matrix that can be loaded with Seurat::Read10X or a txt file. Maybe this is better
handled out of the function since we'd also be dealing with h5 files.}

\item{libraryID}{A character indicating name of the sample.}

\item{studyID}{A character indicating name of the study.}

\item{tissueSource}{A character indicating arterial bed of the library.}

\item{seqPlatform}{A character indicating the sequencing platform.}

\item{UMI}{A boolean indicating whether the read type is a UMI.}

\item{dblFindIter}{A numeric indicating how many iterations we need to run for getting consensus doublets.}

\item{seuratFilter}{A boolean indicating whether we wish to filter low quality cells. 
Set to FALSE in pre-processing round for doublet removal.}

\item{computeSil}{A boolean indicating whether silhouette coefficients should be computed to determine
the resolution parameter for clustering.}

\item{minRes}{A numeric indicating lower boundary for silhouette analysis.}

\item{maxRes}{A numeric indicating upper boundary for silhouette analysis.}

\item{clusteringAlg}{A character vector indicating whether to perform louvain or leiden clustering.}

\item{renameCells}{A boolean indicating if cells should be renamed by appending the library ID as prefix.}

\item{makeAnnData}{A boolean indicating whether we want to generate an .h5ad file for the processed Seurat object.}

\item{annDataParentDir}{A character vector indicating the parent directory where we want to save the produced .h5ad object.}

\item{queryFeatures}{A character vector with genes of interest for checking expression before and after QC.}
}
\value{
A list where the first object is the processed Seurat obj and lots of other QC stats and plots.
}
\description{
A function that will handle QC (doublet and ambient RNA removal) and robust processing for individual scRNA libraries.
}
