% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_processing_pipeline.R
\name{doItAll}
\alias{doItAll}
\title{Process scRNA libraries}
\usage{
doItAll(
  countsMatrix,
  libraryID,
  studyID,
  arterialOrigin,
  minRes = 0.2,
  maxRes = 1.9,
  dblFindIter = 3,
  seuratFilter = TRUE,
  setAutoThreshold = TRUE,
  regressMitoRibo = FALSE,
  rmMitoRiboVarGenes = TRUE,
  autoSelectPCs = TRUE,
  clusteringAlg = "louvain",
  queryFeatures = NULL,
  minUMI = 500,
  maxUMI = 20000,
  minFeatures = 200,
  maxFeatures = 4000,
  maxMtPercent = 10,
  maxHbPercent = 1,
  diseaseStatus = NULL,
  age = NULL,
  sex = NULL,
  race = NULL,
  makeAnnData = FALSE,
  annDataParentDir = NULL
)
}
\arguments{
\item{countsMatrix}{sparse count matrix required as initial input. As of now, make it accept 
a filtered_feature_bc_matrix with a Seurat::Read10X or a txt file. Maybe this is better
handled out of the function since we'd also be dealing with h5 files. Maybe we can do this later
but not priority right now}

\item{libraryID}{A character indicating the id for the library.}

\item{studyID}{A character string indicating the study that generated the data.}

\item{arterialOrigin}{A character vector indicating vascualr bed of the data.}

\item{minRes}{A numeric indicating lower boundary for sil analysis}

\item{maxRes}{A numeric indicating upper boundary for sil analysis}

\item{dblFindIter}{A numeric indicating how many iterations we need to run for getting consensus doublets}

\item{seuratFilter}{A boolean indicating whether we wish to remove low quality cells}

\item{setAutoThreshold}{A boolean indicating whether we should perform adaptive QC thresholding using the MAD.}

\item{regressMitoRibo}{A boolean indicating whether mitochondorial and ribosomal variance should be regressed out.}

\item{rmMitoRiboVarGenes}{A boolean indicating whether to remove mitochondorial and ribosomal genes from set of variable features.}

\item{autoSelectPCs}{A boolean indicating whether to determine the number of PCs in a data-driven way.}

\item{clusteringAlg}{A character vector indicating whether to perform louvain or leiden clustering.}

\item{queryFeatures}{A character vector with genes of interest for checking expression before and after QC.}

\item{minUMI}{A numeric indicating min number of UMIs to keep.}

\item{maxUMI}{A numeric indicating max number of UMIs to keep.}

\item{minFeatures}{A numeric indicating the minimum number of genes cells should express to be included.}

\item{maxFeatures}{A numeric indicating the maximum number of genes cells should express to be included.}

\item{maxMtPercent}{A numeric indicating the max percentage of reads mapped to the mito genome.}

\item{maxHbPercent}{A numeric indicating the max percentage of reads mapped to hemoglobin genes.}

\item{diseaseStatus}{A character vector indicating disease status of the library (e.g., lesion vs non-lesion)}

\item{age}{A character vector indicating age of the subject.}

\item{sex}{A character vector indicating sex of the subject.}

\item{race}{A character vector indicating race of the subject.}

\item{makeAnnData}{A boolean indicating whether we want to generate an .h5ad file for the processed Seurat object.}

\item{annDataParentDir}{A character vector indicating the parent directory where we want to save the produced .h5ad object.}
}
\value{
A list where the first object is the processed Seurat obj and a bunch of other QC stats and plots.
}
\description{
A function that will handle QC (doublet and ambient RNA removal) and robust processing for individual scRNA libraries
}
