% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_processing_pipeline.R
\name{doItAll}
\alias{doItAll}
\title{Process scRNA libraries}
\usage{
doItAll(
  countsMatrix,
  libraryID,
  studyID,
  minRes = 0.2,
  maxRes = 1.9,
  dblFindIter = 3,
  setAutoThreshold = FALSE,
  regressMitoRibo = FALSE,
  rmMitoRiboVarGenes = FALSE,
  clusteringAlg = "louvain",
  queryGenes = NULL,
  minUMI = 500,
  maxUMI = 20000,
  minFeatures = 200,
  maxFeatures = 4000,
  maxMtPercent = 10,
  maxHbPercent = 1,
  arterialOrigin = NULL,
  diseaseStatus = NULL,
  sex = NULL
)
}
\arguments{
\item{countsMatrix}{sparse count matrix required as initial input. As of now, make it accept 
a filtered_feature_bc_matrix with a Seurat::Read10X or a txt file. Maybe this is better
handled out of the function since we'd also be dealing with h5 files. Maybe we can do this later
but not priority right now}

\item{libraryID}{A character indicating the id for the library.}

\item{studyID}{A character string indicating the study that generated the data.}

\item{minRes}{A numeric indicating lower boundary for sil analysis}

\item{maxRes}{A numeric indicating upper boundary for sil analysis}

\item{dblFindIter}{A numeric indicating how many iterations we need to run for getting consensus doublets}

\item{queryGenes}{A character vector with genes of interest for checking expression before and after QC.}

\item{minUMI}{A numeric indicating min number of UMIs to keep.}

\item{maxUMI}{A numeric indicating max number of UMIs to keep.}

\item{minFeatures}{A numeric indicating the minimum number of genes cells should express to be included.}

\item{maxFeatures}{A numeric indicating the maximum number of genes cells should express to be included.}

\item{maxMtPercent}{A numeric indicating the max percentage of reads mapped to the mito genome.}

\item{maxHbPercent}{A numeric indicating the max percentage of reads mapped to hemoglobin genes.}

\item{arterialOrigin}{A character string indicating vascualr bed of the data.}

\item{diseaseStatus}{A character string indicating disease status of the library (e.g., lesion vs non-lesion)}

\item{sex}{A character string indicating sex of the subject.}
}
\value{
A list where the first object is the processed Seurat obj and remaining objects are UMAP plots before and after QC.
}
\description{
A function that will handle QC (doublet and ambient RNA removal) and processing for individual scRNA libraries
}
\examples{
seurat_outs = doitall(counts_matrix, "rpe004", "pan_et_al")
}
