% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_processing_pipeline.R
\name{doitall}
\alias{doitall}
\title{A function that will handle QC (doublet and ambient RNA removal) and processing for individual scRNA libraries}
\usage{
doitall(
  counts_matrix,
  library_id,
  study_id,
  min_res = 0.3,
  max_res = 1.9,
  dbl_remove_iter = 3
)
}
\arguments{
\item{library_id}{A character indicating the id for the library}

\item{min_res}{A numeric indicating lower boundary for sil analysis}

\item{max_res}{A numeric indicating upper boundary for sil analysis}

\item{dbl_remove_iter}{A numeric indicating how many iterations we need to run for getting consensus doublets}

\item{count_matrix}{sparse count matrix required as initial input. As of now, make it accept 
a filtered_feature_bc_matrix with a Seurat::Read10X or a txt file. Maybe this is better
handled out of the function since we'd also be dealing with h5 files. Maybe we can do this later
but not priority right now}
}
\value{
A list where the first object is the processed Seurat obj and remaining objects are UMAP plots before and after QC.
}
\description{
A function that will handle QC (doublet and ambient RNA removal) and processing for individual scRNA libraries
}
\examples{
seurat_outs = doitall(counts_matrix, "rpe004", "pan_et_al")
}
