% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotRNAcomplexity}
\alias{plotRNAcomplexity}
\title{Internal helper function to plot scRNA library complexity}
\usage{
plotRNAcomplexity(seuratObj, cor = FALSE, UMI = TRUE)
}
\arguments{
\item{seuratObj}{A Seurat object with QC metrics stored in the metadata}

\item{cor}{A boolean indicating whether we want to apply a lm to 
UMIs and genes.}

\item{UMI}{A boolean indicating whether the read is a UMI or a raw Read.}

\item{df}{A dataframe with metadata variables nCount_RNA, nFeature_RNA
and percent.mt variables.}
}
\value{
A ggplot object showing the correlation between 
number of UMIs/Reads and detected genes.
}
\description{
this helper function will correlate number of UMIs to
the number of genes detected for each cell, and color them
by the percent of reads aligned to the mitochondrial genome
}
