% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{scDblFinderClusters}
\alias{scDblFinderClusters}
\title{Identify and remove doublets.}
\usage{
scDblFinderClusters(
  seuratObj,
  nrep,
  multisampleDataset = FALSE,
  sampleIDs = NULL
)
}
\arguments{
\item{nrep}{A Numeric indicating the number of times to run the scDblFinder function.}

\item{multisampleDataset}{A Logical indicating whether the processed dataset has multiple samples}

\item{seuratObj:}{A Pre-processed seurat object with cell clusters annotations.}

\item{sample_ids}{if multisample_dataset = TRUE, A character that refers to the metadata column
that contains sample ids.}
}
\value{
A vector of consensus doublet cell barcodes
}
\description{
Internal helper function to automate doublet removal workflow and update metadata.
}
