% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{seuratSCTprocess}
\alias{seuratSCTprocess}
\title{Downstream processing of counts matrix}
\usage{
seuratSCTprocess(
  seuratObj,
  libraryID,
  studyID,
  tissueSource,
  nMADs = 3,
  minReads = 500,
  maxReads = 20000,
  minFeatures = 200,
  maxFeatures = 4000,
  minMtPercent = 0.1,
  maxMtPercent = 20,
  minRiboPercent = 0.1,
  maxHbPercent = 1,
  diseaseStatus = NULL,
  age = NULL,
  sex = NULL,
  ancestry = NULL,
  seqPlatform = "10x",
  seqWorkflow = "cell",
  regressMitoRibo = FALSE,
  rmMitoRiboVarGenes = FALSE,
  seuratFilter = FALSE,
  setAutoThreshold = TRUE,
  autoSelectPCs = TRUE
)
}
\arguments{
\item{seuratObj}{An input Seurat object.}

\item{libraryID}{A character indicating name of the sample.}

\item{studyID}{A character indicating name of the study.}

\item{tissueSource}{A character indicating arterial bed of the library.}

\item{nMADs}{A character indicating the number of Median Absolute Deviations to be used during joint metric filtering 
including nUMIs, nGenes, percent Mito and percent Ribo.}

\item{minReads}{A numeric indicating min number of UMIs to keep.}

\item{maxReads}{A numeric indicating max number of UMIs to keep.}

\item{minFeatures}{A numeric indicating the minimum number of genes cells should express to be included.}

\item{maxFeatures}{A numeric indicating the maximum number of genes cells should express to be included.}

\item{minMtPercent}{A numeric indicating the min percentage of reads mapped to the mito genome.}

\item{maxMtPercent}{A numeric indicating the max percentage of reads mapped to the mito genome.}

\item{minRiboPercent}{A numeric indicating the min percentage of reads mapped to ribosomal genes.}

\item{maxHbPercent}{A numeric indicating the max percentage of reads mapped to hemoglobin genes.}

\item{diseaseStatus}{A character indicating the disease status of the library (e.g., non-lesion, lesion).}

\item{age}{A character vector indicating age of the subject.}

\item{sex}{A character vector indicating sex of the subject.}

\item{seqPlatform}{A character indicating the sequencing platform.}

\item{seqWorkflow}{A character indicating the sequencing workflow, one of cell or nucleus.}

\item{regressMitoRibo}{A boolean indicating whether to regress out Mitochondrial variance or not.}

\item{rmMitoRiboVarGenes}{A boolean indicating whether we should remove Mito and Ribo genes from 
the set of highly variable features.}

\item{seuratFilter}{A boolean indicating whether we wish to filter low quality cells. 
Set to FALSE in pre-processing round for doublet removal.}

\item{setAutoThreshold}{A boolean indicating whether we should do adative thresholding for removing
low quality cells. This approach uses the Median Absolute Deviation (MAD).}

\item{autoSelectPCs}{A boolean indicating whether to determine the number of PCs in a data-driven way.}

\item{race}{A character vector indicating race of the subject.}
}
\value{
A seurat object that has been SCT normalized, nearest neighbors graph and UMAP embeddings.
}
\description{
Internal helper function to standardize filtering, SCT normalization
and dimensionality reduction. This function will also add
some important metadata variables.
}
